Share the optimistic archive cache edit and listing seed with hosts (ERMAIN-614)
