+++ frontend/src/library/index.ts
@@ -23,7 +23,10 @@ export {
   useDelegatedRunHeader,
   type DelegatedRunHeaderState,
 } from "@/hooks/chat/useDelegatedRunHeader";
-export { useGenerationStatusStore } from "@/hooks/chat/store/generationStatusStore";
+export {
+  seedGenerationStatusFromListing,
+  useGenerationStatusStore,
+} from "@/hooks/chat/store/generationStatusStore";
 export { DelegatedRunOpenProvider } from "@/providers/DelegatedRunOpenProvider";
 export {
   ChatMessage,
